Refactor : Image 포맷 관련 커스텀 예외 클래스 제거
- 비즈니스보다 파일 형식과 관련된 예외로 기본 클래스로 변경
- IllegalArgumentException.class
