+++ src/main/java/org/prgrms/wumo/global/exception/GlobalExceptionHandler.java
@@ -1,6 +1,8 @@
 package org.prgrms.wumo.global.exception;
 
 import org.prgrms.wumo.global.exception.custom.DuplicateException;
+import org.prgrms.wumo.global.exception.custom.ImageDeleteFailedException;
+import org.prgrms.wumo.global.exception.custom.ImageUploadFailedException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -13,7 +15,7 @@ public class GlobalExceptionHandler {
 
 	@ExceptionHandler({
 			DuplicateException.class,
-			ImageUploadFailedException.class, InvalidImageFormatException.class, ImageDeleteFailedException.class
+			ImageUploadFailedException.class, IllegalArgumentException.class, ImageDeleteFailedException.class
 	})
 	public ResponseEntity<ExceptionResponse> handleException(RuntimeException runtimeException) {
 		log.info("exception : " + runtimeException);
