--- src/main/java/org/prgrms/wumo/global/exception/InvalidImageFormatException.java
@@ -1,9 +1,0 @@
-package org.prgrms.wumo.global.exception;
-
-public class InvalidImageFormatException extends RuntimeException {
-
-	public InvalidImageFormatException(String message) {
-		super(message);
-	}
-
-}
