+++ src/main/java/org/prgrms/wumo/global/exception/custom/ImageUploadFailedException.java
@@ -1,4 +1,4 @@
-package org.prgrms.wumo.global.exception;
+package org.prgrms.wumo.global.exception.custom;
 
 public class ImageUploadFailedException extends RuntimeException {
